Edit driver database objects implemented

+++ RacedayStopwatch/Drivers/AddDriverTableViewController.swift
@@ -10,10 +10,12 @@ import UIKit
 
 class AddDriverTableViewController: UITableViewController {
     
-
+    
     @IBOutlet weak var driverName: UITextField!
     @IBOutlet weak var driverNumber: UITextField!
     @IBOutlet weak var helmetPickerView: UIPickerView!
+    @IBOutlet weak var saveButton: UIButton!
+    
     var helmets: [UIImage] = [
         UIImage(named: "helmet_red")!,
         UIImage(named: "helmet_blue")!,
@@ -22,48 +24,62 @@ class AddDriverTableViewController: UITableViewController {
         UIImage(named: "helmet_green")!,
         ]
     var driver: Driver?
-
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         driverNumber.keyboardType = .numberPad
-        print("i container\(driver!.name)")
-
+        if let driver = driver {
+            driverName.text = driver.name
+            driverNumber.text = driver.number
+            
+            saveButton.setTitle("Save", for: .normal)
+        }
+        
         // Uncomment the following line to preserve selection between presentations
         // self.clearsSelectionOnViewWillAppear = false
-
+        
         // Uncomment the following line to display an Edit button in the navigation bar for this view controller.
         // self.navigationItem.rightBarButtonItem = self.editButtonItem
     }
-
+    
     // MARK: - Table view data source
-
+    
     override func numberOfSections(in tableView: UITableView) -> Int {
         return 1
     }
-
+    
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return 3
     }
-
-    @IBAction func addDriver(_ sender: Any) {
-        let driver = Driver(context: CoreDataService.context)
-        if let newDriverName = driverName.text, !newDriverName.isEmpty, newDriverName != ""{
-            driver.name = newDriverName
-        }else {
-            driverName.attributedPlaceholder = NSAttributedString(string: "ENTER DRIVER'S NAME HERE",attributes: [NSAttributedString.Key.foregroundColor: UIColor.red])
-        }
-        if let newDriverNumber = driverNumber.text, !newDriverNumber.isEmpty, newDriverNumber != ""{
-            driver.number = newDriverNumber
-        }else{
-            driverNumber.attributedPlaceholder = NSAttributedString(string: "ENTER DRIVER'S CAR NUMBER HERE",attributes: [NSAttributedString.Key.foregroundColor: UIColor.red])
-        }
-        if (driver.name.count > 0 && driver.number.count > 0){
-            print("\(driver.name) - \(driver.name.count) && \(driver.number) - \(driver.number.count)")
+    
+    @IBAction func save(_ sender: Any) {
+        if let driver = driver{
+            print("trying to save an edited driver")
+            driver.name = driverName.text!
+            driver.number = driverNumber.text!
             driver.image = helmets[helmetPickerView.selectedRow(inComponent: 0)].pngData()! as NSData
             CoreDataService.saveContext()
             dismiss(animated: true, completion: nil)
-        }else{
-            CoreDataService.context.delete(driver)
+        }else {
+            let driver = Driver(context: CoreDataService.context)
+            if let newDriverName = driverName.text, !newDriverName.isEmpty, newDriverName != ""{
+                driver.name = newDriverName
+            }else {
+                driverName.attributedPlaceholder = NSAttributedString(string: "ENTER NAME",attributes: [NSAttributedString.Key.foregroundColor: UIColor.red])
+            }
+            if let newDriverNumber = driverNumber.text, !newDriverNumber.isEmpty, newDriverNumber != ""{
+                driver.number = newDriverNumber
+            }else{
+                driverNumber.attributedPlaceholder = NSAttributedString(string: "ENTER CAR NUMBER",attributes: [NSAttributedString.Key.foregroundColor: UIColor.red])
+            }
+            if (driver.name.count > 0 && driver.number.count > 0){
+                print("\(driver.name) - \(driver.name.count) && \(driver.number) - \(driver.number.count)")
+                driver.image = helmets[helmetPickerView.selectedRow(inComponent: 0)].pngData()! as NSData
+                CoreDataService.saveContext()
+                dismiss(animated: true, completion: nil)
+            }else{
+                CoreDataService.context.delete(driver)
+            }
         }
     }
     
@@ -78,7 +94,7 @@ extension AddDriverTableViewController: UIPickerViewDelegate, UIPickerViewDataSo
     }
     
     func pickerView(_ pickerView: UIPickerView, numberOfRowsInComponent component: Int) -> Int {
-        return 5
+        return helmets.count
     }
     
     func pickerView(_ pickerView: UIPickerView, rowHeightForComponent component: Int) -> CGFloat {
